% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxjob64Win.R
\name{pxjob}
\alias{pxjob}
\title{Run PxJob on a px file and save it}
\usage{
pxjob(input, output)
}
\arguments{
\item{input}{Path to input file.}

\item{output}{Path to save parsed px file at.}
}
\value{
PxJob's exit code.
}
\description{
Run PxJob on a raw px file and save the parsed output. PxJob only works on 
Windows.
}
\examples{
pxjob(file.path('px_example.px'), file.path('px_parsed.px'))
}
